+++ bash/run/run_chatbot.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+. .env
+. .env_CHATBOT
+
+python  src/web_chat.py